added a moveInfo method to the gui

+++ Game/src/battleMoves/SingleSelectedHealComponent.java
@@ -28,7 +28,7 @@ public class SingleSelectedHealComponent extends MoveComponent {
 	}
 
 
-	public float getHeal() {
+	public int getHeal() {
 		return this.heal;
 	}
